requirements.txt
streamlit==1.28.1
supabase==2.0.2
openai==1.3.7
langchain==0.1.0
langchain-text-splitters==0.0.1
langchain-community==0.0.10
tiktoken==0.5.2
python-dotenv==1.0.0
pypdf2==3.0.1
markdown==3.5.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0 